Re-do the logic of service and repository classes
