+++ src/main/java/repository/ProductRepository.java
@@ -1,67 +1,30 @@
 package repository;
 
 import jakarta.persistence.EntityManager;
-import jakarta.persistence.EntityManagerFactory;
-import jakarta.persistence.Persistence;
 import model.Product;
 
 import java.util.List;
 
 public class ProductRepository {
-    private final EntityManagerFactory emf;
 
-    public ProductRepository() {
-        emf = Persistence.createEntityManagerFactory("jpaUnit");
-    }
-
-    public void save(Product product) {
-        EntityManager em = emf.createEntityManager();
-        em.getTransaction().begin();
-        em.persist(product);
-        em.getTransaction().commit();
-        em.close();
-    }
-
-    public List<Product> findAll() {
-        EntityManager em = emf.createEntityManager();
+    public List<Product> findAll(EntityManager em) {
         List<Product> products = em.createQuery("SELECT p FROM Product p", Product.class).getResultList();
         em.close();
         return products;
     }
 
-    public Product findById(String id) {
-        EntityManager em = emf.createEntityManager();
+    public Product findById(String id, EntityManager em) {
         Product product = em.find(Product.class, id);
         em.close();
         return product;
     }
 
-    public List<Product> findByCategory(String category) {
-        EntityManager em = emf.createEntityManager();
+    public List<Product> findByCategory(String category, EntityManager em) {
         List<Product> products = em.createQuery("SELECT p FROM Product p WHERE p.categoria = :category", Product.class)
                 .setParameter("category", category)
                 .getResultList();
         em.close();
         return products;
-    }
-
-    public void update(Product product) {
-        EntityManager em = emf.createEntityManager();
-        em.getTransaction().begin();
-        em.merge(product);
-        em.getTransaction().commit();
-        em.close();
-    }
-
-    public void delete(int id) {
-        EntityManager em = emf.createEntityManager();
-        em.getTransaction().begin();
-        Product product = em.find(Product.class, id);
-        if (product != null) {
-            em.remove(product);
-        }
-        em.getTransaction().commit();
-        em.close();
     }
